refactor(types): Move some types from itinerary-body to types package.
