+++ packages/itinerary-body/src/TransitLegBody/index.tsx
@@ -1,7 +1,12 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore FIXME: Create TypeScript types for core-utils packages.
 import coreUtils from "@opentripplanner/core-utils";
-import { Fare, Leg, TransitOperator } from "@opentripplanner/types";
+import {
+  Fare,
+  Leg,
+  LegIconComponent,
+  TransitOperator
+} from "@opentripplanner/types";
 import React, { Component, FunctionComponent, ReactElement } from "react";
 import {
   FormattedMessage,
@@ -13,7 +18,6 @@ import { VelocityTransitionGroup } from "velocity-react";
 
 import * as S from "../styled";
 import {
-  LegIconComponent,
   RouteDescriptionProps,
   SetActiveLegFunction,
   SetViewedTripFunction,
